+++ src/main/java/address/controller/RootLayoutController.java
@@ -9,7 +9,7 @@ import address.exceptions.DuplicateDataException;
 import address.exceptions.DuplicateTagException;
 import address.model.datatypes.tag.Tag;
 import address.model.ModelManager;
-import address.shortcuts.ShortcutsManager;
+import address.keybindings.KeyBindingsManager;
 import address.prefs.PrefsManager;
 
 import com.google.common.eventbus.Subscribe;
@@ -62,11 +62,11 @@ public class RootLayoutController {
         this.mainApp = mainApp;
     }
 
-    public void setShortcuts(){
-        menuFileNew.setAccelerator(ShortcutsManager.BINDINGS.ACCELERATOR_FILE_NEW.getKeyCombination());
-        menuFileOpen.setAccelerator(ShortcutsManager.BINDINGS.ACCELERATOR_FILE_OPEN.getKeyCombination());
-        menuFileSave.setAccelerator(ShortcutsManager.BINDINGS.ACCELERATOR_FILE_SAVE.getKeyCombination());
-        menuFileSaveAs.setAccelerator(ShortcutsManager.BINDINGS.ACCELERATOR_FILE_SAVE_AS.getKeyCombination());
+    public void setAccelerators(){
+        menuFileNew.setAccelerator(KeyBindingsManager.BINDINGS.ACCELERATOR_FILE_NEW.getKeyCombination());
+        menuFileOpen.setAccelerator(KeyBindingsManager.BINDINGS.ACCELERATOR_FILE_OPEN.getKeyCombination());
+        menuFileSave.setAccelerator(KeyBindingsManager.BINDINGS.ACCELERATOR_FILE_SAVE.getKeyCombination());
+        menuFileSaveAs.setAccelerator(KeyBindingsManager.BINDINGS.ACCELERATOR_FILE_SAVE_AS.getKeyCombination());
     }
 
     @Subscribe
