+++ src/main/java/address/keybindings/KeySequence.java
@@ -1,4 +1,4 @@
-package address.shortcuts;
+package address.keybindings;
 
 import address.events.BaseEvent;
 import javafx.scene.input.KeyCode;
